Easier to just move the functions into main, that way I can pass around local variables and such
