--- PipeRec.h
@@ -1,9 +1,0 @@
-//
-// Created by sage7 on 3/19/25.
-//
-
-#ifndef PIPEREC_H
-#define PIPEREC_H
-void parseRadarData();
-void readData();
-#endif //PIPEREC_H
